+++ src/Quartz/Configuration/JobFactoryOptions.cs
@@ -1,5 +1,0 @@
-namespace Quartz;
-
-public class JobFactoryOptions
-{
-}